move helpers.js to helpers/base.js, browserpass/browserpass-extension#320
