+++ src/helpers/ui.js
@@ -4,7 +4,7 @@
 const m = require("mithril");
 const dialog = require("../popup/modalDialog");
 const notify = require("../popup/notifications");
-const helpers = require("../helpers");
+const helpers = require("../helpers/base");
 
 const containsNumbersRegEx = RegExp(/[0-9]/);
 const containsSymbolsRegEx = RegExp(/[\p{P}\p{S}]/, "u");
