+++ src/popup/detailsInterface.js
@@ -2,7 +2,7 @@ module.exports = DetailsInterface;
 
 const m = require("mithril");
 const Moment = require("moment");
-const helpers = require("../helpers");
+const helpers = require("../helpers/base");
 const helpersUI = require("../helpers/ui");
 const layout = require("./layoutInterface");
 const Login = require("./models/Login");
