+++ src/popup/models/Login.js
@@ -2,7 +2,7 @@
 
 require("chrome-extension-async");
 const sha1 = require("sha1");
-const helpers = require("../../helpers");
+const helpers = require("../../helpers/base");
 const helpersUI = require("../../helpers/ui");
 const Settings = require("./Settings");
